move parameter-related fields logic from project-base to the framework and frontend-api packages
