+++ packages/framework/src/Model/Product/Search/ProductElasticsearchConverter.php
@@ -99,6 +99,11 @@ class ProductElasticsearchConverter
             $filledParameter['parameter_value_id'] = $parameter['parameter_value_id'] ?? '';
             $filledParameter['parameter_value_uuid'] = $parameter['parameter_value_uuid'] ?? '';
             $filledParameter['parameter_value_text'] = $parameter['parameter_value_text'] ?? '';
+            $filledParameter['parameter_group'] = $parameter['parameter_group'] ?? null;
+            $filledParameter['parameter_unit'] = $parameter['parameter_unit'] ?? null;
+            $filledParameter['parameter_value_for_slider_filter'] = $parameter['parameter_value_for_slider_filter'] ?? null;
+            $filledParameter['ordering_priority'] = $parameter['ordering_priority'] ?? null;
+            $filledParameter['parameter_type'] = $parameter['parameter_type'] ?? null;
             $results[] = $filledParameter;
         }
