+++ packages/framework/tests/Unit/Model/Product/Search/ProductElasticsearchConverterTest.php
@@ -111,6 +111,10 @@ class ProductElasticsearchConverterTest extends TestCase
                     'parameter_name' => '',
                     'parameter_value_uuid' => '',
                     'parameter_value_text' => '',
+                    'parameter_group' => null,
+                    'parameter_unit' => null,
+                    'parameter_value_for_slider_filter' => null,
+                    'parameter_type' => null,
                 ],
             ],
             'availability' => '',
