+++ project-base/app/src/FrontendApi/Resolver/Products/DataMapper/ProductEntityFieldMapper.php
@@ -11,7 +11,6 @@ use App\Model\Product\Product;
 use App\Model\Product\ProductRepository;
 use GraphQL\Executor\Promise\Promise;
 use Overblog\DataLoader\DataLoaderInterface;
-use Override;
 use Shopsys\FrameworkBundle\Component\Breadcrumb\BreadcrumbFacade;
 use Shopsys\FrameworkBundle\Component\Domain\Domain;
 use Shopsys\FrameworkBundle\Component\Router\FriendlyUrl\FriendlyUrlFacade;
@@ -20,7 +19,6 @@ use Shopsys\FrameworkBundle\Model\Pricing\Group\PricingGroupSettingFacade;
 use Shopsys\FrameworkBundle\Model\Product\Accessory\ProductAccessoryFacade;
 use Shopsys\FrameworkBundle\Model\Product\Availability\ProductAvailabilityFacade;
 use Shopsys\FrameworkBundle\Model\Product\Collection\ProductCollectionFacade;
-use Shopsys\FrameworkBundle\Model\Product\Product as BaseProduct;
 use Shopsys\FrameworkBundle\Model\Product\ProductFrontendLimitProvider;
 use Shopsys\FrameworkBundle\Model\Product\ProductVisibilityFacade;
 use Shopsys\FrameworkBundle\Model\ProductVideo\ProductVideoTranslationsRepository;
@@ -62,6 +60,9 @@ use Shopsys\FrontendApiBundle\Model\Resolver\Products\DataMapper\ProductEntityFi
  * @method bool isAllowedNegativeStock(\App\Model\Product\Product $product)
  * @method bool isSellingDenied(\App\Model\Product\Product $product)
  * @method bool isCurrentlyOutOfStock(\App\Model\Product\Product $product)
+ * @property \App\Model\Product\ProductRepository $productRepository
+ * @property \App\Model\Product\Parameter\ParameterRepository $parameterRepository
+ * @method \Shopsys\FrontendApiBundle\Model\Parameter\ParameterWithValues[] getParameters(\App\Model\Product\Product $product)
  */
 class ProductEntityFieldMapper extends BaseProductEntityFieldMapper
 {
@@ -82,8 +83,8 @@ class ProductEntityFieldMapper extends BaseProductEntityFieldMapper
      * @param \Shopsys\FrameworkBundle\Component\Router\FriendlyUrl\FriendlyUrlFacade $friendlyUrlFacade
      * @param \Shopsys\FrameworkBundle\Model\Stock\ProductStockFacade $productStockFacade
      * @param \App\Model\Product\ProductRepository $productRepository
-     * @param \Shopsys\FrameworkBundle\Model\Pricing\Group\PricingGroupSettingFacade $pricingGroupSettingFacade
      * @param \App\Model\Product\Parameter\ParameterRepository $parameterRepository
+     * @param \Shopsys\FrameworkBundle\Model\Pricing\Group\PricingGroupSettingFacade $pricingGroupSettingFacade
      * @param \Shopsys\FrameworkBundle\Component\Breadcrumb\BreadcrumbFacade $breadcrumbFacade
      * @param \Overblog\DataLoader\DataLoaderInterface $categoriesBatchLoader
      * @param \Overblog\DataLoader\DataLoaderInterface $brandsBatchLoader
@@ -104,9 +105,9 @@ class ProductEntityFieldMapper extends BaseProductEntityFieldMapper
         ProductVideoTranslationsRepository $productVideoTranslationsRepository,
         FriendlyUrlFacade $friendlyUrlFacade,
         ProductStockFacade $productStockFacade,
-        protected readonly ProductRepository $productRepository,
+        ProductRepository $productRepository,
+        ParameterRepository $parameterRepository,
         protected readonly PricingGroupSettingFacade $pricingGroupSettingFacade,
-        protected readonly ParameterRepository $parameterRepository,
         protected readonly BreadcrumbFacade $breadcrumbFacade,
         protected readonly DataLoaderInterface $categoriesBatchLoader,
         protected readonly DataLoaderInterface $brandsBatchLoader,
@@ -127,6 +128,8 @@ class ProductEntityFieldMapper extends BaseProductEntityFieldMapper
             $productVideoTranslationsRepository,
             $friendlyUrlFacade,
             $productStockFacade,
+            $productRepository,
+            $parameterRepository,
         );
     }
 
@@ -191,31 +194,6 @@ class ProductEntityFieldMapper extends BaseProductEntityFieldMapper
         ksort($flagsIndexedById);
 
         return array_values($flagsIndexedById);
-    }
-
-    /**
-     * Method is overridden, so it returns parameters for the variants too.
-     *
-     * @param \App\Model\Product\Product $product
-     * @return array
-     */
-    #[Override]
-    public function getParameters(BaseProduct $product): array
-    {
-        $products = [];
-
-        if ($product->isMainVariant() === true) {
-            $products = $this->productRepository->getAllSellableVariantsByMainVariant(
-                $product,
-                $this->domain->getId(),
-                $this->currentCustomerUser->getPricingGroup(),
-            );
-        }
-        $products[] = $product;
-
-        $productParameterValuesData = $this->parameterRepository->getProductParameterValuesDataByProducts($products, $this->domain->getLocale());
-
-        return $this->parameterWithValuesFactory->createParametersArrayFromProductArray(['parameters' => $productParameterValuesData]);
     }
 
     /**
