+++ project-base/app/src/Model/Product/Parameter/ParameterRepository.php
@@ -4,10 +4,8 @@ declare(strict_types=1);
 
 namespace App\Model\Product\Parameter;
 
-use Doctrine\ORM\Query\Expr\Join;
 use Override;
 use Shopsys\FrameworkBundle\Model\Product\Parameter\ParameterRepository as BaseParameterRepository;
-use Shopsys\FrameworkBundle\Model\Product\Parameter\ProductParameterValue;
 
 /**
  * @method \Shopsys\FrameworkBundle\Model\Product\Parameter\Parameter[] getParametersUsedByProductsInCategory(\App\Model\Category\Category $category, \Shopsys\FrameworkBundle\Component\Domain\Config\DomainConfig $domainConfig)
@@ -20,56 +18,10 @@ use Shopsys\FrameworkBundle\Model\Product\Parameter\ProductParameterValue;
  * @method \Shopsys\FrameworkBundle\Model\Product\Parameter\Parameter[] getParametersUsedByProductsInCategoryWithoutSlider(\App\Model\Category\Category $category, int $domainId)
  * @method \Shopsys\FrameworkBundle\Model\Product\Parameter\ParameterValue[] getParameterValuesUsedByProductsInCategoryByParameter(\App\Model\Category\Category $category, \Shopsys\FrameworkBundle\Model\Product\Parameter\Parameter $parameter, int $domainId, string $locale)
  * @method \Doctrine\ORM\QueryBuilder getProductParameterValuesByProductSortedByOrderingPriorityAndNameQueryBuilder(\App\Model\Product\Product $product, string $locale)
+ * @method array getProductParameterValuesDataByProducts(\App\Model\Product\Product[] $products, string $locale)
  */
 class ParameterRepository extends BaseParameterRepository
 {
-    /**
-     * @param \App\Model\Product\Product[] $products
-     * @param string $locale
-     * @return array
-     */
-    public function getProductParameterValuesDataByProducts(array $products, string $locale): array
-    {
-        if (count($products) === 0) {
-            return [];
-        }
-
-        return $this->em->createQueryBuilder()
-            ->select(
-                'p.id as parameter_id,
-                p.orderingPriority as ordering_priority,
-                p.parameterType as parameter_type,
-                pv.id as parameter_value_id,
-                p.uuid as parameter_uuid,
-                pt.name as parameter_name,
-                pv.uuid as parameter_value_uuid,
-                pv.text as parameter_value_text,
-                pv.numericValue as parameter_value_numeric_value,
-                pgt.name as parameter_group,
-                pg.position as group_position,
-                put.name as parameter_unit',
-            )
-            ->distinct()
-            ->from(ProductParameterValue::class, 'ppv')
-            ->join('ppv.parameter', 'p')
-            ->join('p.translations', 'pt', Join::WITH, 'pt.locale = :locale AND pt.name IS NOT NULL')
-            ->leftjoin('p.group', 'pg')
-            ->leftJoin('pg.translations', 'pgt', Join::WITH, 'pgt.locale = :locale AND pgt.name IS NOT NULL')
-            ->leftJoin('p.unit', 'pu')
-            ->leftJoin('pu.translations', 'put', Join::WITH, 'put.locale = :locale AND put.name IS NOT NULL')
-            ->join('ppv.value', 'pv', Join::WITH, 'pv.locale = :locale')
-            ->where('ppv.product IN (:products)')
-            ->orderBy('group_position', 'ASC')
-            ->addOrderBy('ordering_priority', 'DESC')
-            ->addOrderBy('parameter_name', 'ASC')
-            ->setParameters([
-                'products' => $products,
-                'locale' => $locale,
-            ])
-            ->getQuery()
-            ->execute();
-    }
-
     /**
      * @param array $productIdsAndParameterNamesAndValues
      * @return string[][]
